remove pkg/helpers/helpers.go
